[WIP] delete all old functions
